run linters
